canary_api/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib_metadata==8.5.0
Markdown==3.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
zipp==3.20.2
